requirements.txt
# ============================================================================
# IoT Bluetooth Network - Requirements
# Projeto SIC - Rede ad-hoc IoT baseada em Bluetooth BLE
# ============================================================================

# ----------------------------------------------------------------------------
# BLE (Bluetooth Low Energy)
# ----------------------------------------------------------------------------
# SimpleBLE - Cliente BLE (recomendado no projeto)
# Nota: Pode requerer instalação de dependências do sistema
# sudo apt-get install cmake build-essential libdbus-1-dev
simpleble>=0.0.5

# D-Bus - Para GATT Server (interface com BlueZ)
dbus-python>=1.3.2

# PyGObject - GLib mainloop para D-Bus
PyGObject>=3.42.0

# ----------------------------------------------------------------------------
# Segurança (50% da avaliação!)
# ----------------------------------------------------------------------------
# Cryptography - X.509, ECDSA (P-521), ECDH, HMAC
cryptography>=41.0.0

# PyDTLS - DTLS para comunicação end-to-end
# Nota: Se tiver problemas, podemos implementar DTLS manualmente com cryptography
pyDTLS>=1.3.0; platform_system != "Windows"

# ----------------------------------------------------------------------------
# Networking & Serialização
# ----------------------------------------------------------------------------
# Scapy - Manipulação de pacotes (opcional, se necessário)
# scapy>=2.5.0

# ----------------------------------------------------------------------------
# CLI & Interface
# ----------------------------------------------------------------------------
# Typer - CLI moderna
typer>=0.9.0

# Rich - UI no terminal (tabelas, progress bars, status)
rich>=13.7.0

# Click - Dependência do Typer
click>=8.1.0

# ----------------------------------------------------------------------------
# Logging
# ----------------------------------------------------------------------------
# Loguru - Sistema de logging avançado
loguru>=0.7.0

# ----------------------------------------------------------------------------
# Configuração
# ----------------------------------------------------------------------------
# Python-dotenv - Gestão de variáveis de ambiente
python-dotenv>=1.0.0

# ----------------------------------------------------------------------------
# Utilidades
# ----------------------------------------------------------------------------
# Threading e Async (built-in no Python 3.7+)
# - asyncio
# - threading
# - queue
# - struct
# - uuid
# - datetime

# ----------------------------------------------------------------------------
# Desenvolvimento (opcional)
# ----------------------------------------------------------------------------
# pytest - Testes unitários
pytest>=7.4.0

# pytest-asyncio - Suporte para testes assíncronos
pytest-asyncio>=0.21.0

# black - Code formatting
black>=23.0.0

# flake8 - Linting
flake8>=6.0.0

# mypy - Type checking
mypy>=1.5.0
